restructure testing of queries
