+++ api/api/src/main/scala/uk/ac/wellcome/platform/api/QueryParams.scala
@@ -175,7 +175,9 @@ object MultipleWorksParams extends QueryParamsUtils {
   implicit val _queryTypeDecoder: Decoder[SearchQueryType] =
     decodeOneWithDefaultOf(
       SearchQueryType.default,
-      "FixedFields" -> SearchQueryType.FixedFields
+      "ConstScore" -> SearchQueryType.ConstScore,
+      "BoolBoosted" -> SearchQueryType.BoolBoosted,
+      "Core" -> SearchQueryType.Core
     )
 }
