+++ api/api/src/main/scala/uk/ac/wellcome/platform/api/services/ElasticsearchService.scala
@@ -64,7 +64,7 @@ class ElasticsearchService(elasticClient: ElasticClient)(
       subType = "elastic",
       action = "query")({
 
-      val searchRequest = ElasticSearchRequestBuilder(
+      val searchRequest = ElasticsearchRequestBuilder(
         index,
         sortDefinitions,
         queryOptions
