+++ api/api/src/main/scala/uk/ac/wellcome/platform/api/swagger/SwaggerDocs.scala
@@ -265,7 +265,7 @@ trait MultipleWorksSwagger {
           "Which query should we use search the works? Used predominantly for internal testing of relevancy. Considered Unstable.",
         schema = new Schema(
           `type` = "enum",
-          allowableValues = Array("IdSearch", "FixedFields"),
+          allowableValues = Array("ConstScore", "BoolBoosted"),
         ),
         required = false
       )
