+++ api/api/src/test/scala/uk/ac/wellcome/platform/api/services/ElasticsearchServiceTest.scala
@@ -20,7 +20,6 @@ import uk.ac.wellcome.models.work.internal.WorkType.{
 }
 import uk.ac.wellcome.models.work.internal._
 import uk.ac.wellcome.platform.api.generators.SearchOptionsGenerators
-import uk.ac.wellcome.platform.api.models.SearchQueryType.IdSearch
 import uk.ac.wellcome.platform.api.models.{
   ItemLocationTypeFilter,
   SearchQuery,
@@ -83,7 +82,7 @@ class ElasticsearchServiceTest
               .queryResults(
                 index,
                 createElasticsearchQueryOptionsWith(
-                  searchQuery = Some(SearchQuery("A", IdSearch))))
+                  searchQuery = Some(SearchQuery("A"))))
 
             whenReady(searchResponseFuture) { response =>
               searchResponseToWorks(response) shouldBe works
@@ -242,273 +241,6 @@ class ElasticsearchServiceTest
                 locationTypeIds = List("iiif-image", "digit")))
             ),
             expectedWorks = List(work, work2)
-          )
-        }
-      }
-    }
-
-    describe("ID search with SearchQuery") {
-      it("searches the canonicalId") {
-        withLocalWorksIndex { index =>
-          val work = createIdentifiedWorkWith(
-            canonicalId = "abc123"
-          )
-
-          insertIntoElasticsearch(index, work)
-
-          assertSearchResultsAreCorrect(
-            index = index,
-            queryOptions = createElasticsearchQueryOptionsWith(
-              searchQuery = Some(SearchQuery("abc123", IdSearch))),
-            expectedWorks = List(work)
-          )
-        }
-      }
-
-      it("searches the sourceIdentifiers") {
-        withLocalWorksIndex { index =>
-          val work = createIdentifiedWorkWith(
-            canonicalId = "abc123",
-            sourceIdentifier = createSourceIdentifierWith()
-          )
-          val workNotMatching = createIdentifiedWorkWith(
-            canonicalId = "123abc",
-            sourceIdentifier = createSourceIdentifierWith()
-          )
-          val query = work.sourceIdentifier.value
-
-          insertIntoElasticsearch(index, work, workNotMatching)
-          assertSearchResultsAreCorrect(
-            index = index,
-            queryOptions = createElasticsearchQueryOptionsWith(
-              searchQuery = Some(SearchQuery(query, IdSearch))),
-            expectedWorks = List(work)
-          )
-        }
-      }
-
-      it("searches the otherIdentifiers") {
-        withLocalWorksIndex { index =>
-          val work = createIdentifiedWorkWith(
-            canonicalId = "abc123",
-            otherIdentifiers = List(createSourceIdentifierWith())
-          )
-          val workNotMatching = createIdentifiedWorkWith(
-            canonicalId = "123abc",
-            otherIdentifiers = List(createSourceIdentifierWith())
-          )
-          val query = work.otherIdentifiers.head.value
-
-          insertIntoElasticsearch(index, work, workNotMatching)
-
-          assertSearchResultsAreCorrect(
-            index = index,
-            queryOptions = createElasticsearchQueryOptionsWith(
-              searchQuery = Some(SearchQuery(query, IdSearch))),
-            expectedWorks = List(work)
-          )
-        }
-      }
-
-      it("searches the items.canonicalId as keyword") {
-        withLocalWorksIndex { index =>
-          val work = createIdentifiedWorkWith(
-            canonicalId = "abc123",
-            items = List(createIdentifiedItemWith(canonicalId = "def"))
-          )
-          val workNotMatching = createIdentifiedWorkWith(
-            canonicalId = "123abc",
-            items = List(createIdentifiedItemWith(canonicalId = "def456"))
-          )
-          val query = "def"
-
-          insertIntoElasticsearch(index, work, workNotMatching)
-
-          assertSearchResultsAreCorrect(
-            index = index,
-            queryOptions = createElasticsearchQueryOptionsWith(
-              searchQuery = Some(SearchQuery(query, IdSearch))),
-            expectedWorks = List(work)
-          )
-        }
-      }
-
-      it("searches the items.sourceIdentifiers") {
-        withLocalWorksIndex { index =>
-          val work = createIdentifiedWorkWith(
-            canonicalId = "abc123",
-            items = List(
-              createIdentifiedItemWith(sourceIdentifier =
-                createSourceIdentifierWith(value = "sourceIdentifier123")))
-          )
-          val workNotMatching = createIdentifiedWorkWith(
-            canonicalId = "123abc",
-            items = List(
-              createIdentifiedItemWith(sourceIdentifier =
-                createSourceIdentifierWith(value = "sourceIdentifier456")))
-          )
-
-          val query = "sourceIdentifier123"
-
-          insertIntoElasticsearch(index, work, workNotMatching)
-
-          assertSearchResultsAreCorrect(
-            index = index,
-            queryOptions = createElasticsearchQueryOptionsWith(
-              searchQuery = Some(SearchQuery(query, IdSearch))),
-            expectedWorks = List(work)
-          )
-        }
-      }
-
-      it("searches the items.otherIdentifiers") {
-        withLocalWorksIndex { index =>
-          val work = createIdentifiedWorkWith(
-            canonicalId = "abc123",
-            items = List(createIdentifiedItemWith(otherIdentifiers =
-              List(createSourceIdentifierWith(value = "sourceIdentifier123"))))
-          )
-          val workNotMatching = createIdentifiedWorkWith(
-            canonicalId = "def456",
-            items = List(createIdentifiedItemWith(otherIdentifiers =
-              List(createSourceIdentifierWith(value = "sourceIdentifier456"))))
-          )
-          val query = "sourceIdentifier123"
-
-          insertIntoElasticsearch(index, work, workNotMatching)
-
-          assertSearchResultsAreCorrect(
-            index = index,
-            queryOptions = createElasticsearchQueryOptionsWith(
-              searchQuery = Some(SearchQuery(query, IdSearch))),
-            expectedWorks = List(work)
-          )
-        }
-      }
-
-      it("matches when searching for an ID") {
-        withLocalWorksIndex { index =>
-          val work = createIdentifiedWorkWith(
-            canonicalId = "abc123"
-          )
-          val query = "abc123"
-
-          insertIntoElasticsearch(index, work)
-
-          assertSearchResultsAreCorrect(
-            index = index,
-            queryOptions = createElasticsearchQueryOptionsWith(
-              searchQuery = Some(SearchQuery(query, IdSearch))),
-            expectedWorks = List(work)
-          )
-        }
-      }
-
-      it("matches when searching for multiple IDs") {
-        withLocalWorksIndex { index =>
-          val work1 = createIdentifiedWorkWith(
-            canonicalId = "abc123"
-          )
-          val work2 = createIdentifiedWorkWith(
-            canonicalId = "123abc"
-          )
-          val query = "abc123 123abc"
-
-          insertIntoElasticsearch(index, work1, work2)
-
-          assertSearchResultsAreCorrect(
-            index = index,
-            queryOptions = createElasticsearchQueryOptionsWith(
-              searchQuery = Some(SearchQuery(query, IdSearch))),
-            expectedWorks = List(work1, work2)
-          )
-        }
-      }
-
-      it("doesn't match on partial IDs") {
-        withLocalWorksIndex { index =>
-          val work1 = createIdentifiedWorkWith(
-            canonicalId = "abcdefg"
-          )
-          val work2 = createIdentifiedWorkWith(
-            canonicalId = "1234567"
-          )
-          val query = "123 abcdefg"
-
-          insertIntoElasticsearch(index, work1, work2)
-
-          assertSearchResultsAreCorrect(
-            index = index,
-            queryOptions = createElasticsearchQueryOptionsWith(
-              searchQuery = Some(SearchQuery(query, IdSearch))),
-            expectedWorks = List(work1)
-          )
-        }
-      }
-
-      it("matches IDs case insensitively") {
-        withLocalWorksIndex { index =>
-          val work1 = createIdentifiedWorkWith(
-            canonicalId = "AbCDeF1234"
-          )
-          val work2 = createIdentifiedWorkWith(
-            canonicalId = "12345Ef"
-          )
-          val query = "abcdef1234 12345ef"
-
-          insertIntoElasticsearch(index, work1, work2)
-
-          assertSearchResultsAreCorrect(
-            index = index,
-            queryOptions = createElasticsearchQueryOptionsWith(
-              searchQuery = Some(SearchQuery(query, IdSearch))),
-            expectedWorks = List(work1, work2)
-          )
-        }
-      }
-
-      it("matches if there is extra terms in the query") {
-        withLocalWorksIndex { index =>
-          val work1 = createIdentifiedWorkWith(
-            canonicalId = "AbCDeF1234"
-          )
-          val work2 = createIdentifiedWorkWith(
-            canonicalId = "12345Ef"
-          )
-          val query = "abcdef1234 12345ef hats, dogs and dolomites"
-
-          insertIntoElasticsearch(index, work1, work2)
-
-          assertSearchResultsAreCorrect(
-            index = index,
-            queryOptions = createElasticsearchQueryOptionsWith(
-              searchQuery = Some(SearchQuery(query, IdSearch))),
-            expectedWorks = List(work1, work2)
-          )
-        }
-      }
-
-      it("puts ID matches at the top of the results") {
-        withLocalWorksIndex { index =>
-          val workWithMatchingTitle = createIdentifiedWorkWith(
-            title = Some("Standing on wrong side of horse")
-          )
-          val workWithMatchingId = createIdentifiedWorkWith(
-            canonicalId = "AbCDeF1234"
-          )
-
-          val query = "abcdef1234 Standing on wrong side of horse"
-
-          insertIntoElasticsearch(
-            index,
-            workWithMatchingTitle,
-            workWithMatchingId)
-
-          assertSearchResultsAreCorrectAndInOrder(
-            index = index,
-            queryOptions = createElasticsearchQueryOptionsWith(
-              searchQuery = Some(SearchQuery(query, IdSearch))),
-            expectedWorks = List(workWithMatchingId, workWithMatchingTitle)
           )
         }
       }
@@ -749,13 +481,6 @@ class ElasticsearchServiceTest
     queryOptions: ElasticsearchQueryOptions,
     expectedWorks: List[IdentifiedWork]) = {
     searchResults(index, queryOptions) should contain theSameElementsAs expectedWorks
-  }
-
-  private def assertSearchResultsAreCorrectAndInOrder(
-    index: Index,
-    queryOptions: ElasticsearchQueryOptions,
-    expectedWorks: List[IdentifiedWork]) = {
-    searchResults(index, queryOptions) should be(expectedWorks)
   }
 
   private def assertListResultsAreCorrect(
